requirements.txt
lit
vaderSentiment==3.3.2
wordcloud==1.9.3
matplotlib==3.9.0
requests==2.32.3
beautifulsoup4==4.12.3
regex==2024.5.15
nltk
numpy==1.26.4
scikit-learn==1.5.0
Flask==3.0.3
flask
torch==2.3.1
PyAudio==0.2.14
Flask-SQLAlchemy==3.1.1
flask_cors
bert-extractive-summarizer
deepmultilingualpunctuation
tensorflow
tf-keras==2.16.0